Move Manager into one file
There are two files defining the Manager type. Move all declarations in
runner.go into manager.go so development doesn't need to change files to
find all the methods of a Manager.
